packs/net-zero/PACK-026-sme-net-zero/requirements.txt
# PACK-026: SME Net Zero Pack - Production Requirements
# Python 3.11+

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Async & HTTP
httpx==0.26.0
aiohttp==3.9.1
asyncio==3.4.3

# Database
psycopg[binary,pool]==3.1.16
psycopg-pool==3.2.0
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlrd==2.0.1

# Calculations & Math
decimal==1.0.0
sympy==1.12
scipy==1.11.4

# Date/Time
python-dateutil==2.8.2
pytz==2023.3.post1

# File Handling
pyyaml==6.0.1
toml==0.10.2
python-multipart==0.0.6

# PDF Processing
pypdf2==3.0.1
reportlab==4.0.8
pillow==10.2.0

# OAuth & Authentication
authlib==1.3.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4

# API Clients (Accounting Software)
xero-python==5.3.0
intuit-oauth==1.2.4
requests-oauthlib==1.3.1

# Validation & Security
email-validator==2.1.0.post1
python-dotenv==1.0.0
bcrypt==4.1.2

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-psycopg==0.43b0
opentelemetry-exporter-prometheus==0.43b0

# Logging
structlog==24.1.0
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0
hypothesis==6.96.1

# Code Quality
black==23.12.1
ruff==0.1.13
mypy==1.8.0
types-pyyaml==6.0.12.12
types-requests==2.31.0.20240106

# Development
ipython==8.19.0
ipdb==0.13.13

# GreenLang Core Dependencies
# (Assumes GreenLang core is installed in the environment)
# If running standalone, uncomment and adjust path:
# -e ../../../greenlang
